MAINT: move basic docstrings and rename to `uarray`
