+++ scipy/fft/_fftlog_multimethods.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 
-from ._basic_np import _dispatch
+from ._basic_uarray import _dispatch
 from ._fftlog_np import fht as _fht
 from ._fftlog_np import ifht as _ifht
 from scipy._lib.uarray import Dispatchable
